save and backup. spiout and beeper tasks seem to work. Startup HardFault appears to be tied to stack usage printout in the default task, but maybe not

+++ Ourtasks/BeepTask.h
@@ -17,9 +17,9 @@
 
 struct BEEPQ
 {
-	uint32_t duron;   // Duration (timer ticks) beep is ON
-	uint32_t duroff;  // Duration (timer ticks) beep is OFF (before finished)
-	uint32_t repct;   // Repetition ct of this beep
+	uint16_t duron;   // Duration (timer ticks) beep is ON
+	uint16_t duroff;  // Duration (timer ticks) beep is OFF (before finished)
+	uint16_t repct;   // Repetition ct of this beep
 };
 
 /* *************************************************************************/
